+++ Example/SwiftyTesseractRTE/ViewController.swift
@@ -26,22 +26,66 @@ class ViewController: UIViewController {
   
   var engine: SwiftyTesseractRTE!
   var excludeLayer: CAShapeLayer!
+  var flashlightButton: UIBarButtonItem!
+  var recognitionButton: UIButton!
+  var recognitionTitleLabel: UILabel!
+  var recognitionLabel: UILabel!
   
   @IBOutlet weak var informationLabel: UILabel!
   @IBOutlet weak var previewView: UIView!
   @IBOutlet weak var regionOfInterest: UIView!
-  @IBOutlet weak var flashLightButton: UIButton!
-  @IBOutlet weak var recognitionButton: UIButton!
-  @IBOutlet weak var recognitionTextView: UITextView!
   @IBOutlet weak var regionOfInterestWidth: NSLayoutConstraint!
   @IBOutlet weak var regionOfInterestHeight: NSLayoutConstraint!
   
   override func viewDidLoad() {
     super.viewDidLoad()
-    let swiftyTesseract = SwiftyTesseract(language: .english)
-    engine = SwiftyTesseractRTE(swiftyTesseract: swiftyTesseract, desiredReliability: .verifiable)
-    engine.recognitionIsActive = false
-    engine.delegate = self
+    
+    // MARK: - UI Setup
+    
+    let navigationBar = navigationController?.navigationBar
+    let textAttributes = [NSAttributedStringKey.foregroundColor: UIColor.white]
+    
+    if #available(iOS 11.0, *) {
+      navigationBar?.prefersLargeTitles = true
+      navigationBar?.largeTitleTextAttributes = textAttributes
+    }
+    
+    navigationBar?.titleTextAttributes = textAttributes
+    navigationBar?.isTranslucent = false
+    navigationBar?.barTintColor = .black
+    navigationItem.title = "SwiftyTesseractRTE"
+    
+    flashlightButton = UIBarButtonItem(title: "Flashlight On", style: .plain, target: self, action: #selector(flashLightButtonTapped(_:)))
+    navigationItem.rightBarButtonItem = flashlightButton
+    
+    recognitionButton = UIButton()
+    recognitionButton.setTitleColor(view.tintColor, for: .normal)
+    recognitionButton.setTitle("Start Recognition", for: .normal)
+    recognitionButton.addTarget(self, action: #selector(recognitionButtonTapped(_:)), for: .touchUpInside)
+    
+    recognitionTitleLabel = UILabel()
+    recognitionTitleLabel.text = "Recognition Text:"
+    recognitionTitleLabel.font = UIFont.systemFont(ofSize: 16, weight: .bold)
+    recognitionTitleLabel.textColor = .white
+    
+    recognitionLabel = UILabel()
+    recognitionLabel.textAlignment = .center
+    recognitionLabel.numberOfLines = 20
+    recognitionLabel.textColor = .white
+    recognitionLabel.text = "Let's Do This!"
+    
+    let stackView = UIStackView(arrangedSubviews: [recognitionButton, recognitionTitleLabel, recognitionLabel])
+    stackView.axis = .vertical
+    stackView.alignment = .center
+    stackView.distribution = .fill
+    stackView.spacing = 8.0
+    stackView.translatesAutoresizingMaskIntoConstraints = false
+    
+    view.addSubview(stackView)
+    stackView.widthAnchor.constraint(equalTo: view.widthAnchor).isActive = true
+    stackView.centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
+    stackView.topAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
+    stackView.bottomAnchor.constraint(lessThanOrEqualTo: view.bottomAnchor).isActive = true
     
     let panGesture = UIPanGestureRecognizer(target: self, action: #selector(handlePan(_:)))
     
@@ -50,12 +94,17 @@ class ViewController: UIViewController {
     regionOfInterest.layer.borderColor = UIColor.blue.cgColor
     regionOfInterest.backgroundColor = .clear
     
-    recognitionButton.setTitle("Start Recognition", for: .normal)
-    
     excludeLayer = CAShapeLayer()
     excludeLayer.fillRule = kCAFillRuleEvenOdd
     excludeLayer.fillColor = UIColor.black.cgColor
     excludeLayer.opacity = 0.7
+    
+    // SwiftyTesseractRTE Setup
+    
+    let swiftyTesseract = SwiftyTesseract(language: .english)
+    engine = SwiftyTesseractRTE(swiftyTesseract: swiftyTesseract, desiredReliability: .verifiable)
+    engine.recognitionIsActive = false
+    engine.delegate = self
     
     engine.startPreview()
   }
@@ -103,12 +152,11 @@ class ViewController: UIViewController {
     informationLabel.isHidden = true
   }
   
-  
-  @IBAction func recognitionButtonTapped(_ sender: Any) {
+  @objc func recognitionButtonTapped(_ sender: Any) {
     recognitionIsRunning.toggle()
   }
   
-  @IBAction func flashLightButtonTapped(_ sender: Any) {
+  @objc func flashLightButtonTapped(_ sender: Any) {
     guard let device = AVCaptureDevice.default(for: .video) else { return }
     if device.hasTorch {
       do {
@@ -119,8 +167,8 @@ class ViewController: UIViewController {
         print("Error: \(e.localizedDescription)")
       }
     }
-    let flashlightImage = device.torchMode == .off ? UIImage(named: "flashlightOff") : UIImage(named: "flashlightOn")
-    flashLightButton.setImage(flashlightImage, for: .normal)
+    let flashlightButtonTitle = device.torchMode == .off ? "Flashlight On" : "Flashlight Off"
+    flashlightButton.title = flashlightButtonTitle
   }
 }
 
@@ -129,7 +177,7 @@ extension ViewController: SwiftyTesseractRTEDelegate {
   func onRecognitionComplete(_ recognizedString: String) {
     AudioServicesPlayAlertSound(SystemSoundID(kSystemSoundID_Vibrate))
     DispatchQueue.main.async { [weak self] in
-      self?.recognitionTextView.text = recognizedString
+      self?.recognitionLabel.text = recognizedString
     }
     recognitionIsRunning = false
   }
@@ -137,8 +185,10 @@ extension ViewController: SwiftyTesseractRTEDelegate {
 }
 
 extension Bool {
+  
   mutating func toggle() {
     self = !self
   }
+  
 }