+++ SwiftyTesseractRTE/Classes/Protocols/AVManager.swift
@@ -8,10 +8,21 @@
 
 import AVFoundation
 
-protocol AVManager: class {
+/// To be implemented if custom AVFoundation logic is needed. This may be desireable if using
+/// SwiftyTesseractRTE in tandem with another class or library that requires CMSampleBuffers
+/// for processing.
+public protocol AVManager: class {
+  /// The `AVCaptureVideoPreviewLayer` displayed to the user
   var previewLayer: AVCaptureVideoPreviewLayer { get }
+  /// The quality of the previewLayer video session.
   var cameraQuality: AVCaptureSession.Preset { get set }
+  /// The underlying capture session providing previewLayer it's video feed
   var captureSession: AVCaptureSession { get }
 
+  /// The delegate to receive the `CMSampleBuffer`s for processing.
+  /// *Note*: If you are creating a custom object to conform to AVManager, you must
+  /// manually set SwiftyTesseractRTE as it's delegate for SwiftyTesseractRTE to
+  /// receive the `CMSampleBuffer`s, otherwise they will not be received to be
+  /// processed for OCR
   var delegate: AVCaptureVideoDataOutputSampleBufferDelegate? { get set }
 }